cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vastctl-cli"
version = "0.1.0"
description = "CLI for Vast.ai GPU instance management (third-party)"
readme = "README.md"
license = "MIT"
authors = [
    { name = "VastCtl Contributors" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Environment :: Console",
]
requires-python = ">=3.8"
dependencies = [
    "vastctl-core==0.1.0",
    "click>=8.0",
    "rich>=12.0",
    "humanize>=4.0",
    "pyyaml>=6.0",
]

[project.scripts]
vastctl = "vastctl_cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vastctl_cli"]
